fix: Update webpack-dev-server config
These new configs makes the html generated by django look for
the static files on the dev-server, only in the production enviroment.
Doing that only by changing the pubicPath prop into webpack.config
creates a problem with sockjs-node. Because of that I had to instruct the
dev-server to point to django server.
So the project now should be ran into the port 3000.